my homework 39-41 lesson

+++ src/main/java/com/tms/service/UserService.java
@@ -5,6 +5,7 @@ import com.tms.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -21,20 +22,31 @@ public class UserService {
         return userRepository.getUserById(id);
     }
 
-    public Optional<User> updateUser(User user){
-        Boolean result = userRepository.updateUser(user);
-        if(result){
-            return getUserById(user.getId());
+    public List<User> getAllUsers() {
+        return userRepository.findAllUsers();
+    }
+    public Optional<User> updateUser(User user) {
+        Boolean updated = userRepository.updateUser(user);
+        if (updated) {
+            return userRepository.getUserById(user.getId());
         }
         return Optional.empty();
     }
 
-    public Optional<User> deleteUser(Long id){
-        Boolean result = userRepository.deleteUser(id);
-        if(result){
-            return getUserById(id);
+    public Optional<User> deleteUser(Long id) {
+        Optional<User> userOptional = userRepository.getUserById(id);
+        System.out.println("Starting the process of deleting a user with ID:"+ id);
+        if (userOptional.isEmpty()) {
+            System.out.println("User with ID not found: "+ id);
+            return Optional.empty();
         }
-        return Optional.empty();
+        User user = userOptional.get();
+        user.setDeleted(true);
+        Boolean isDeleted = userRepository.deleteUser(id);
+        if (!isDeleted) {
+            return Optional.empty();
+        }
+        return Optional.of(user);
     }
     
     public Optional<User> createUser(User user){
